[CodeNarc] Update Gradle wrapper to 6.9.2
Gradle 6.5 uses vulnerable versions of log4j. Gradle 6.9.2 is the last
update the Gradle 6 line, and the only Gradle 6 that is fixed for
log4shell.

Moving to Gradle 7 was not possible due to broken backwards
compatibility. Maybe later, if the base version used for CodeNarc in
this project is updated.