[CodeNarc] Suggest non-automated install-file method
The automated method via Maven's install-plugin only worked on my
machine on condition that CodeNarc was already in the local repository.
Otherwise, the build failed before reaching that step.

This method has the advantage that it won't try to build the project
when installing CodeNarc to the local repository, and it also won't
deploy it on each build, saving what should be an unnecessary step as
long as CodeNarc hasn't been modified.