[CodeNarc] Automate CodeNarc build from Maven's initialize
`mvn initialize` will now call the gradle build for CodeNarc, thus
reducing the number of necessary steps to start working on the project.

The initialize can be run on the root project. The only side-effect is
that it will start two jacoco-maven-plugin:prepare-agent. Those don't
affect the build, but they could be avoided by changing their execution
phase (maybe a phase that would execute it only if tests are run?).

.gradle has been added to the root .gitignore (since the command is not
run from the same place and results in a .gradle folder that is not
ignored), and comments and a new rule have been added.